adapted some classes to work with the Google AppEngine

+++ org.cumulus4j.store/src/main/java/org/cumulus4j/store/model/EncryptionCoordinateSet.java
@@ -139,7 +139,7 @@ public class EncryptionCoordinateSet
 
 	@PrimaryKey
 	@Persistent(valueStrategy=IdGeneratorStrategy.NATIVE, sequence="EncryptionCoordinateSetSequence")
-	private int encryptionCoordinateSetID = -1;
+	private long encryptionCoordinateSetID = -1;
 
 	@Persistent(nullValue=NullValue.EXCEPTION)
 	private String cipherTransformation;
@@ -187,7 +187,7 @@ public class EncryptionCoordinateSet
 	 * @return the unique numeric identifier (primary key).
 	 */
 	public int getEncryptionCoordinateSetID() {
-		return encryptionCoordinateSetID;
+		return (int)encryptionCoordinateSetID;
 	}
 
 	/**
